+++ ihmc-avatar-interfaces/src/test/java/us/ihmc/avatar/AvatarPauseWalkingTest.java
@@ -183,7 +183,7 @@ public abstract class AvatarPauseWalkingTest implements MultiRobotTestInterface
       FootstepDataMessage footstepData = new FootstepDataMessage();
       footstepData.setLocation(stepLocation);
       footstepData.setOrientation(orient);
-      footstepData.setRobotSide(robotSide);
+      footstepData.setRobotSide(robotSide.toByte());
       message.add(footstepData);
    }
