+++ ihmc-avatar-interfaces/src/test/java/us/ihmc/avatar/circleWalkTest/HumanoidCircleWalkTest.java
@@ -218,7 +218,7 @@ public abstract class HumanoidCircleWalkTest implements MultiRobotTestInterface
       FootstepDataMessage footstepData = new FootstepDataMessage();
       footstepData.setLocation(stepLocation);
       footstepData.setOrientation(orient);
-      footstepData.setRobotSide(robotSide);
+      footstepData.setRobotSide(robotSide.toByte());
       message.add(footstepData);
    }
