+++ ihmc-common-walking-control-modules/src/test/java/us/ihmc/commonWalkingControlModules/controlModules/rigidBody/RigidBodyControlManagerTest.java
@@ -297,7 +297,7 @@ public class RigidBodyControlManagerTest
 
          SE3TrajectoryMessage message = HumanoidMessageTools.createSE3TrajectoryMessage(1, worldFrame);
          message.setTrajectoryPoint(0, trajectoryTime, position, orientation, linearVelocity, angularVelocity, worldFrame);
-         message.getQueueingProperties().setExecutionMode(ExecutionMode.OVERRIDE);
+         message.getQueueingProperties().setExecutionMode(ExecutionMode.OVERRIDE.toByte());
          message.getQueueingProperties().setPreviousMessageId((long) -1);
 
          SelectionMatrix6D selectionMatrix6D = new SelectionMatrix6D();
